View virtualization
InspireView scaled up to VDOMView
InspireView and PerspireView extends from VDOMView
created global.name and global.self proxies
